refactor: 옵션 반환 API들 위치 수정
Trim과 Engine에 존재하던 옵션 반환 API들 Option으로 이동
