+++ BE-MyCarMaster/src/test/java/softeer/be_my_car_master/api/option/controller/DefaultOptionControllerTest.java
@@ -1,6 +1,5 @@
 package softeer.be_my_car_master.api.option.controller;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -24,6 +23,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import softeer.be_my_car_master.api.option.dto.response.DefaultOptionDto;
 import softeer.be_my_car_master.api.option.dto.response.GetDefaultOptionsResponse;
 import softeer.be_my_car_master.api.option.usecase.GetDefaultOptionsUseCase;
+import softeer.be_my_car_master.api.trim.dto.response.GetTrimDefaultOptionsResponse;
+import softeer.be_my_car_master.api.trim.dto.response.TrimDefaultOptionDto;
+import softeer.be_my_car_master.api.trim.usecase.GetTrimDefaultOptionsUseCase;
 import softeer.be_my_car_master.domain.option.Category;
 import softeer.be_my_car_master.global.response.Response;
 import softeer.be_my_car_master.global.response.ResponseStatus;
@@ -39,6 +41,45 @@ class DefaultOptionControllerTest {
 
 	@MockBean
 	private GetDefaultOptionsUseCase getDefaultOptionsUseCase;
+	@MockBean
+	private GetTrimDefaultOptionsUseCase getTrimDefaultOptionsUseCase;
+
+	@Nested
+	@DisplayName("getTrimDefaultOptions Test")
+	class GetTrimDefaultOptionsTest {
+
+		@Test
+		@DisplayName("트림의 기본옵션 목록을 조회합니다")
+		void getTrimDefaultOptions() throws Exception {
+			//given
+			GetTrimDefaultOptionsResponse response = new GetTrimDefaultOptionsResponse();
+			TrimDefaultOptionDto trimDefaultOptionDto = TrimDefaultOptionDto.builder()
+				.id(1L)
+				.category(Category.SAFE.getValue())
+				.name("HTRAC")
+				.imgUrl("imgUrl")
+				.description("옵션 상세설명")
+				.build();
+			response.setDefaultOptions(Arrays.asList(trimDefaultOptionDto));
+
+			given(getTrimDefaultOptionsUseCase.execute(anyLong())).willReturn(response);
+
+			Response successResponse = Response.createSuccessResponse(response);
+			String responseBody = objectMapper.writeValueAsString(successResponse);
+
+			//when
+			ResultActions perform = mockMvc.perform(
+				get("/trims/{trimId}/default-options", 1L)
+					.contentType(MediaType.APPLICATION_FORM_URLENCODED)
+			);
+
+			//then
+			perform
+				.andExpect(status().isOk())
+				.andExpect(content().contentType("application/json"))
+				.andExpect(content().json(responseBody, true));
+		}
+	}
 
 	@Nested
 	@DisplayName("getDefaultOptions Test")
