+++ BE-MyCarMaster/src/test/java/softeer/be_my_car_master/api/option/controller/OptionControllerTest.java
@@ -21,6 +21,9 @@ import org.springframework.test.web.servlet.ResultActions;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import softeer.be_my_car_master.api.engine.dto.response.GetUnselectableOptionsByEngineResponse;
+import softeer.be_my_car_master.api.engine.dto.response.UnselectableOptionDto;
+import softeer.be_my_car_master.api.engine.usecase.GetUnselectableOptionsByEngineUseCase;
 import softeer.be_my_car_master.api.option.dto.response.AppliedOptionDto;
 import softeer.be_my_car_master.api.option.dto.response.DefaultOptionDto;
 import softeer.be_my_car_master.api.option.dto.response.FilterDto;
@@ -48,10 +51,13 @@ class OptionControllerTest {
 
 	@MockBean
 	private GetOptionsUseCase getOptionsUseCase;
+	@MockBean
+	private GetUnselectableOptionsByEngineUseCase getUnselectableOptionsByEngineUseCase;
 
 	@Nested
 	@DisplayName("getOptions Test")
 	class GetOptionsTest {
+
 		@Test
 		@DisplayName("선택 가능한 옵션 목록을 조회합니다")
 		void getOptions() throws Exception {
@@ -329,6 +335,126 @@ class OptionControllerTest {
 		}
 	}
 
+	@Nested
+	@DisplayName("getUnselectableOptionsByEngine Test")
+	class GetUnselectableOptionsByEngineTest {
+
+		@Test
+		@DisplayName("변경하려는 엔진에 따라 선택불가능해지는 옵션 목록을 조회합니다")
+		void getUnselectableOptionsByEngine() throws Exception {
+			//given
+			GetUnselectableOptionsByEngineResponse getUnselectableOptionsByEngineResponse =
+				new GetUnselectableOptionsByEngineResponse();
+			UnselectableOptionDto unselectableOptionDto = UnselectableOptionDto.builder()
+				.id(1L)
+				.name("주차보조 시스템||")
+				.price(1280000)
+				.build();
+			getUnselectableOptionsByEngineResponse.setUnselectableOptions(Arrays.asList(unselectableOptionDto));
+
+			given(getUnselectableOptionsByEngineUseCase.execute(any(), any(), any())).willReturn(
+				getUnselectableOptionsByEngineResponse);
+
+			Response successResponse = Response.createSuccessResponse(getUnselectableOptionsByEngineResponse);
+			String responseBody = objectMapper.writeValueAsString(successResponse);
+
+			//when
+			ResultActions perform = mockMvc.perform(
+				get("/engines/1/unselectable-options")
+					.contentType(MediaType.APPLICATION_FORM_URLENCODED)
+					.param("trimId", "1")
+					.param("optionIds", "1,2")
+			);
+
+			//then
+			perform
+				.andExpect(status().isOk())
+				.andExpect(content().contentType("application/json"))
+				.andExpect(content().json(responseBody, true));
+		}
+
+		@Test
+		@DisplayName("trimId는 1 이상이어야 합니다")
+		void minimumTrimId() throws Exception {
+			//given
+			String responseBody = getClientErrorResponseBody();
+
+			//when
+			ResultActions perform = mockMvc.perform(
+				get("/engines/1/unselectable-options")
+					.contentType(MediaType.APPLICATION_FORM_URLENCODED)
+					.param("trimId", "0")
+					.param("optionIds", "1,2")
+			);
+
+			//then
+			perform
+				.andExpect(status().is4xxClientError())
+				.andExpect(content().contentType("application/json"))
+				.andExpect(content().json(responseBody, false));
+		}
+
+		@Test
+		@DisplayName("trimId는 null값 일 수 없습니다")
+		void nonNullTrimId() throws Exception {
+			//given
+			String responseBody = getClientErrorResponseBody();
+
+			//when
+			ResultActions perform = mockMvc.perform(
+				get("/engines/1/unselectable-options")
+					.contentType(MediaType.APPLICATION_FORM_URLENCODED)
+					.param("optionIds", "1,2")
+			);
+
+			//then
+			perform
+				.andExpect(status().is4xxClientError())
+				.andExpect(content().contentType("application/json"))
+				.andExpect(content().json(responseBody, false));
+		}
+
+		@Test
+		@DisplayName("optionIds는 null값 일 수 없습니다")
+		void nonNullOptionIds() throws Exception {
+			//given
+			String responseBody = getClientErrorResponseBody();
+
+			//when
+			ResultActions perform = mockMvc.perform(
+				get("/engines/1/unselectable-options")
+					.contentType(MediaType.APPLICATION_FORM_URLENCODED)
+					.param("trimId", "1")
+			);
+
+			//then
+			perform
+				.andExpect(status().is4xxClientError())
+				.andExpect(content().contentType("application/json"))
+				.andExpect(content().json(responseBody, false));
+		}
+
+		@Test
+		@DisplayName("optionIds는 빈 List일 수 없습니다")
+		void nonEmptyOptionIds() throws Exception {
+			//given
+			String responseBody = getClientErrorResponseBody();
+
+			//when
+			ResultActions perform = mockMvc.perform(
+				get("/engines/1/unselectable-options")
+					.contentType(MediaType.APPLICATION_FORM_URLENCODED)
+					.param("trimId", "1")
+					.param("optionIds", "")
+			);
+
+			//then
+			perform
+				.andExpect(status().is4xxClientError())
+				.andExpect(content().contentType("application/json"))
+				.andExpect(content().json(responseBody, false));
+		}
+	}
